+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -4,8 +4,8 @@ import com.techcourse.controller.dto.LoginRequest;
 import com.techcourse.db.InMemoryUserRepository;
 import com.techcourse.model.User;
 import java.util.UUID;
-import org.apache.catalina.Session;
-import org.apache.catalina.SessionManager;
+import org.apache.catalina.session.Session;
+import org.apache.catalina.session.SessionManager;
 import org.apache.catalina.controller.AbstractController;
 import org.apache.catalina.util.StaticResourceReader;
 import org.apache.coyote.http11.common.HttpStatusCode;
