+++ tomcat/src/main/java/org/apache/catalina/session/Manager.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.session;
 
 import jakarta.servlet.http.HttpSession;
 import java.io.IOException;
