+++ tomcat/src/main/java/org/apache/catalina/session/Session.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.session;
 
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.http.HttpSession;
