+++ tomcat/src/test/java/com/techcourse/controller/LoginControllerTest.java
@@ -10,11 +10,11 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.util.Map;
 import java.util.UUID;
-import org.apache.catalina.Session;
-import org.apache.catalina.SessionManager;
 import org.apache.catalina.route.DefaultDispatcher;
 import org.apache.catalina.route.RequestMapper;
 import org.apache.catalina.route.RequestMapping;
+import org.apache.catalina.session.Session;
+import org.apache.catalina.session.SessionManager;
 import org.apache.coyote.Dispatcher;
 import org.apache.coyote.http11.Http11Processor;
 import org.junit.jupiter.api.BeforeEach;
